Make the LMFDB pip installable
- Add pyproject.toml (setuptools) with metadata, dependencies, package data
  (templates, static files, yaml) and entry points: after
  `sage -pip install -e .` (which is what requirements.txt now does),
  `import lmfdb` works without PYTHONPATH manipulation and the website can
  be started from anywhere with `lmfdb` or `sage -python -m lmfdb`.
  The version number is single-sourced in lmfdb/version.py.  Since
  psycodict is not yet released on PyPI, it is installed from its git
  repository; that must change before the lmfdb can be released on PyPI.

- Importing lmfdb no longer connects to postgres: lmfdb.db is now a lazy
  proxy that connects on first use (or explicitly via db.connect()), so
  `import lmfdb` is cheap and has no side effects (no config.ini,
  flasklog, slow_queries.log or secret_key created).

- Move the configuration code to lmfdb/config.py so that using the
  database does not pull in lmfdb.utils (and thus flask and sage);
  lmfdb/utils/config.py remains as a backwards-compatible shim, still
  runnable as a script as the CI does.  When running from a git checkout
  everything is located as before (config.ini and secret_key at the root);
  when installed as a package they live in ~/.lmfdb (overridable via
  LMFDB_HOME/LMFDB_CONFIG or --config-file), with a config.ini in the
  current directory taking precedence.

- Log files no longer pollute the current directory: flasklog,
  slow_queries.log and verification logs now default to logs/ under the
  checkout root (or ~/.lmfdb/logs when installed).  Existing config files
  with the old default values are redirected transparently; explicitly
  configured paths are respected.

- Move CONTRIBUTORS.yaml into the lmfdb package so the acknowledgments
  page works when installed, and run the knowl code-reference git greps
  in the checkout root rather than the process working directory (they
  now degrade gracefully when lmfdb is not run from a checkout).
