+++ .gitignore
@@ -14,6 +14,12 @@ config.ini
 
 flasklog
 slow_queries.log
+/logs/
+
+# Packaging artifacts
+*.egg-info/
+build/
+dist/
 
 # Testing files
 htmlcov
