+++ GettingStarted.md
@@ -57,6 +57,15 @@ Installation
    sage -pip install -r requirements.txt
    ```
 
+   This installs the LMFDB itself (in editable mode, i.e. pointing at your
+   checkout) together with its dependencies, so that `import lmfdb` works
+   from `sage` regardless of the current directory; there is no need to add
+   the checkout to your `PYTHONPATH`.  Equivalently, you can run
+   `sage -pip install -e .` (add `[dev]` to also get the testing tools).
+   Importing `lmfdb` does not connect to the database; the connection is
+   established the first time `lmfdb.db` is used, or explicitly via
+   `db.connect()`.
+
    ### Troubleshooting with packages.
 
    - If you have not run the site for a while you might get an error
@@ -129,6 +138,10 @@ Running
    sage -python start-lmfdb.py --debug
    ```
 
+   Since the LMFDB is installed by the step above, you can equivalently run
+   `lmfdb --debug` (a script installed by pip) or
+   `sage -python -m lmfdb --debug`, from any directory.
+
  * The effect of the (optional) --debug is that you will be running
    with the beta flag switched on as at dev.lmfdb.org, and also that
    if code fails your browser will show useful debugging information.
@@ -162,6 +175,17 @@ Running
    on devmirror.lmfdb.xyz) is sufficient, and this step is not necessary. If you do plan
    on using a different database instance, you can do so by changing
    config.ini in the root of the lmfdb directory.
+
+ * Locations of configuration and log files: when running from a git
+   checkout, the configuration file is `config.ini` at the root of the
+   checkout (as it always was) and log files (`flasklog`,
+   `slow_queries.log`, verification logs) go to the `logs/` subdirectory.
+   When the LMFDB is installed as a package rather than run from a checkout,
+   these files live in `~/.lmfdb` instead (unless there is a `config.ini` in
+   the current directory).  The environment variables `LMFDB_HOME` (the
+   directory for all of these files) and `LMFDB_CONFIG` (the path of the
+   configuration file) override these defaults, as does the `--config-file`
+   command-line option.
 
 
 CoCalc
