+++ README.md
@@ -14,7 +14,11 @@ Wiki
 Development
 -----------
 
-* [Getting Started](https://github.com/LMFDB/lmfdb/blob/main/GettingStarted.md) -- cheat sheet for setting up lmfdb
+* [Getting Started](https://github.com/LMFDB/lmfdb/blob/main/GettingStarted.md) -- cheat sheet for setting up lmfdb.
+  The LMFDB is pip installable: `sage -pip install -e .` from a checkout
+  (this is what `sage -pip install -r requirements.txt` does), after which
+  `import lmfdb` works from `sage` anywhere and the website can be started
+  with `lmfdb` or `sage -python -m lmfdb`.
 * [Development Guide](https://github.com/LMFDB/lmfdb/blob/main/Development.md) -- organizing development
 * [Style Guide](https://github.com/LMFDB/lmfdb/blob/main/StyleGuide.md) -- how things on LMFDB pages should be styled to give the web site a coherent look
 * [Command-line search](https://github.com/LMFDB/lmfdb/blob/main/CommandLineSearch.md) -- searching the LMFDB from the command line with `lmfdb_search`
