+++ lmfdb/__main__.py
@@ -1,0 +1,7 @@
+"""
+Start the LMFDB website via ``python -m lmfdb`` (or ``sage -python -m lmfdb``).
+"""
+from lmfdb.website import main
+
+if __name__ == "__main__":
+    main()
