+++ lmfdb/version.py
@@ -1,0 +1,3 @@
+# The single source for the LMFDB version number, used both by
+# lmfdb/app.py and by the packaging metadata in pyproject.toml
+version = "1.2.1"
